chore: export statsblock prop types
